requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
Django==2.0.7
django-crontab==0.7.1
gunicorn==19.9.0
idna==2.7
numpy==1.14.5
pandas==0.23.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.18
pymssql==2.1.3
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.5
pyzabbix==0.7.4
requests==2.19.1
six==1.11.0
urllib3==1.23
